Restructure the application a little bit
