+++ wear/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:required="false" />
 
         <activity
-            android:name=".GameActivity"
+            android:name=".Game.GameActivity"
             android:label="@string/app_name"
             android:launchMode="singleTask"
             android:theme="@android:style/Theme.DeviceDefault.Light">
@@ -39,7 +39,7 @@
         </service>
 
         <activity
-            android:name=".FinishActivity"
+            android:name=".Game.FinishActivity"
             android:label="@string/title_activity_finish"
             android:launchMode="singleTask"
             android:theme="@android:style/Theme.DeviceDefault.Light" />
