--- wear/src/main/java/ca/hoogit/powerhour/GamePagerAdapter.java
@@ -1,58 +1,0 @@
-/**
- * Copyright (C) 2015, Jordon de Hoog
- * <p/>
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- * <p/>
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * <p/>
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- */
-package ca.hoogit.powerhour;
-
-import android.app.Fragment;
-import android.app.FragmentManager;
-import android.support.wearable.view.FragmentGridPagerAdapter;
-
-import java.util.List;
-
-/**
- * @author jordon
- *
- * Date    20/12/15
- * Description
- * Adapter for the pager, contains game screen and controls
- */
-public class GamePagerAdapter extends FragmentGridPagerAdapter {
-
-    private List<Fragment> mFragments;
-    private boolean mPagingEnabled = true;
-
-    public GamePagerAdapter(FragmentManager fm, List<Fragment> fragments) {
-        super(fm);
-        mFragments = fragments;
-    }
-
-    @Override
-    public Fragment getFragment(int row, int column) {
-        return mFragments.get(column);
-    }
-
-    @Override
-    public int getRowCount() {
-        return 1;
-    }
-
-    @Override
-    public int getColumnCount(int row) {
-        return mFragments == null ? 0 : mFragments.size();
-    }
-
-}
