+++ wear/src/main/res/layout/activity_game.xml
@@ -7,5 +7,5 @@
     android:layout_height="match_parent"
     app:rectLayout="@layout/rect_activity_main"
     app:roundLayout="@layout/round_activity_main"
-    tools:context=".GameActivity"
+    tools:context=".Game.GameActivity"
     tools:deviceIds="wear" />
